moved statement test under jdbc package folder (#149)
* moved statement  test under jdbc package folder

* moving items to unit directory
