+++ src/test/java/com/microsoft/sqlserver/jdbc/unit/statement/MergeTest.java
@@ -5,7 +5,7 @@
  * 
  * This program is made available under the terms of the MIT License. See the LICENSE file in the project root for more information.
  */
-package com.microsoft.sqlserver.UnitStatement;
+package com.microsoft.sqlserver.jdbc.unit.statement;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
